+++ modules/tinymce/src/core/main/ts/api/ui/Ui.ts
@@ -19,6 +19,7 @@ export interface EditorUiApi {
   enable: () => void;
   disable: () => void;
   isDisabled: () => boolean;
+  setToolbarOffset: (offset: number) => void;
 }
 
 export interface EditorUi extends EditorUiApi {