+++ modules/tinymce/src/themes/silver/main/ts/ui/header/StaticHeader.ts
@@ -9,10 +9,12 @@ import { Fun } from '@ephox/katamari';
 
 const setup = Fun.noop;
 const isDocked = Fun.never;
+const setToolbarOffset = Fun.noop;
 const getBehaviours = Fun.constant([]);
 
 export {
   setup,
   isDocked,
-  getBehaviours
+  getBehaviours,
+  setToolbarOffset
 };